requirements.txt
Flask
gunicorn
numpy
pandas
tensorflow-cpu
flask-cors
psycopg2-binary
python-dotenv
scikit-learn